pyproject.toml
[project]
name = "smoke-fire-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.2",
    "keras-image-helper>=0.0.2",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "requests>=2.32.5",
    "ruff>=0.14.13",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
